app/image_finder/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
Pillow>=9.0.0
faiss-cpu>=1.7.4
numpy>=1.21.0
pandas>=1.3.0
tqdm>=4.64.0
